Fixed a bug in getZero whereby 0-0 would be zero that wasn't flagged ZERO.
After Googling and reading my old architecture text, I have a better getOverflow behavior (thanks HackerNews), and we now force two's complement semantics.

Added OVERFLOW detection to single-word-result multiply.
Changed multi-word division to flag OVERFLOW instead of INVALID. Possibly fixed a bug in it, too.

Silenced a warning with aggressive optimization in GCC. Well, GCC, if 'may' is ever 'is', then that is a bug in the compiler.